requirements.txt

# FastAPI and web server
fastapi~=0.115.12
uvicorn~=0.34.2
python-dotenv~=1.1.0
python-multipart==0.0.6

# Optional dependencies for API testing
aiohttp~=3.11.18

# Browser automation
playwright==1.38.0

# LLM integration
langchain-anthropic==0.0.9
langchain-openai==0.0.2
langchain-google-genai==0.0.3
langchain-mistralai==0.0.1

# Image processing
Pillow~=11.2.1

# For async operations
asyncio==3.4.3
pydantic~=2.10.6
jwt~=1.3.1
bcrypt~=4.3.0
motor~=3.7.1
pymongo~=4.13.0
cryptography~=44.0.3

reportlab==4.0.7
weasyprint==61.2
Pillow==10.1.0